handle: Introduce pack
A Pack is a wrapper around a static vector, which guarantees that the
vector does not move from its original memory location. Also some helper
methods are available to find an element by its position and to push a
new element into the vector.

Implement clone method which copies the element pointed to by the handle
pushes it into the vector, and returns a new handle to the new element.