handle: Generational indices
Improve pack by introducing another indirection layer with a list of
generational indices which handles point to. Both handles and indices
have a generation counter which should be equal for the handle to be
valid. With this new indirection, elements are guaranteed to be
contiguous in memory with no holes between them.